Add the excel data as zip
More data for interested parties.
Each folder corresponds to when the bluetooth tag was in that location. For example, all of the data when the tags were in the Clean room are in the "Clean" folder.

Each excel sheet in the folder is labelled as Location_Tag#_Timestamp. 

In each excel sheet there is a tag, node_id which corresponds to a zone, the average RSSI signal and normalized RSSI ratio. The x and y positions are also provided for each zone. These positions are for the receiver positions and do not change in the excel sheets.